Fix pagination link to first page
Thanks so much for this theme, I really like the way it looks, and I appreciate all the extra features.

I found that the pagination controls link the number 1 to the previous page instead of the first page, perhaps because you can't just link to `/blog/page1`. I've implemented the correct link by splitting on `/`, popping the `page:num` or whatever pattern is configured, and then rejoining on `/`. I think this is general enough to cover any configuration that the pagination module will support.